feat: add AppPagination named export for app router
Extract the presentational component into `src/pagination.tsx` and inject
a routing adapter from per-router wrappers:
- `PagesPagination` (default export) uses `next/router` + `next/head`
- `AppPagination` uses `next/navigation` and omits the `<link rel=prev|next>`
  SEO hints (client components cannot write to `<head>`)

The default export keeps pointing at `PagesPagination` for backwards
compatibility. Both named exports ship from the single entry point.

Postbuild fixes for microbundle-crl:
- Prepend `'use client'` so Next.js treats the bundled module as a client
  component (the directive is stripped during rollup bundling)
- Inject `__esModule` marker so CJS consumers resolve the default export
  via `.default` instead of getting the whole exports object
