+++ package.json
@@ -12,7 +12,7 @@
     "node": "^20 || ^22 || ^24"
   },
   "scripts": {
-    "build": "microbundle-crl --no-compress --format modern,cjs",
+    "build": "microbundle-crl --no-compress --format modern,cjs && node scripts/add-use-client.js",
     "start": "microbundle-crl watch --no-compress --format modern,cjs",
     "prepare": "run-s build",
     "test": "run-s test:types test:unit",
