+++ src/index.test.ts
@@ -1,10 +1,16 @@
-import Pagination from '.'
+import Pagination, { PagesPagination, AppPagination } from '.'
 import { getSizes } from './lib/sizes'
 import assert from 'assert'
 
 describe('Pagination', () => {
   it('is truthy', () => {
     expect(Pagination).toBeTruthy()
+  })
+  it('default export is PagesPagination', () => {
+    expect(Pagination).toBe(PagesPagination)
+  })
+  it('exports AppPagination', () => {
+    expect(AppPagination).toBeTruthy()
   })
 })
